backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai-whisper
google-generativeai==0.4.1
moviepy==1.0.3
tqdm==4.66.2